Move InventorySlots to requirements package. Add Predicate to ItemSlots checkInventory method for consistency.
